feat: change weight fields to float64

+++ entity/assignment.go
@@ -1,10 +1,10 @@
 package entity
 
 type AssignmentGroup struct {
-	Id       string `json:"id" gorm:"primaryKey;type:char(255)"`
-	Name     string `json:"name"`
-	CourseId string `json:"courseId"`
-	Weight   int    `json:"weight"`
+	Id       string  `json:"id" gorm:"primaryKey;type:char(255)"`
+	Name     string  `json:"name"`
+	CourseId string  `json:"courseId"`
+	Weight   float64 `json:"weight"`
 
 	Assignments []Assignment `gorm:"foreignKey:AssignmentGroupId" json:"assignments,omitempty"`
 
@@ -15,7 +15,7 @@ type Assignment struct {
 	Id                               string                   `json:"id" gorm:"primaryKey;type:char(255)"`
 	Name                             string                   `json:"name"`
 	Description                      string                   `json:"description"`
-	MaxScore                         int                      `json:"maxScore"`
+	MaxScore                         float64                  `json:"maxScore"`
 	ExpectedScorePercentage          float64                  `json:"expectedScorePercentage"`
 	ExpectedPassingStudentPercentage float64                  `json:"expectedPassingStudentPercentage"`
 	IsIncludedInClo                  *bool                    `json:"isIncludedInClo"`
@@ -68,8 +68,8 @@ type AssignmentUseCase interface {
 	GetByCourseId(courseId string) ([]Assignment, error)
 	GetByGroupId(assignmentGroupId string) ([]Assignment, error)
 	GetPassingStudentPercentage(assignmentId string) (float64, error)
-	Create(assignmentGroupId string, name string, description string, maxScore int, expectedScorePercentage float64, expectedPassingStudentPercentage float64, courseLearningOutcomeIds []string, isIncludedInClo bool) error
-	Update(id string, name string, description string, maxScore int, expectedScorePercentage float64, expectedPassingStudentPercentage float64, isIncludedInClo bool) error
+	Create(assignmentGroupId string, name string, description string, maxScore float64, expectedScorePercentage float64, expectedPassingStudentPercentage float64, courseLearningOutcomeIds []string, isIncludedInClo bool) error
+	Update(id string, name string, description string, maxScore float64, expectedScorePercentage float64, expectedPassingStudentPercentage float64, isIncludedInClo bool) error
 	Delete(id string) error
 
 	CreateLinkCourseLearningOutcome(assignmentId string, courseLearningOutcomeId []string) error
@@ -77,7 +77,7 @@ type AssignmentUseCase interface {
 
 	GetGroupByGroupId(assignmentGroupId string) (*AssignmentGroup, error)
 	GetGroupByCourseId(courseId string, withAssignment bool) ([]AssignmentGroup, error)
-	CreateGroup(name string, courseId string, weight int) error
-	UpdateGroup(assignmentGroupId string, name string, weight int) error
+	CreateGroup(name string, courseId string, weight float64) error
+	UpdateGroup(assignmentGroupId string, name string, weight float64) error
 	DeleteGroup(assignmentGroupId string) error
 }
