+++ infrastructure/fiber/request/assignment_group.go
@@ -1,12 +1,12 @@
 package request
 
 type CreateAssignmentGroupPayload struct {
-	Name     string `json:"name" validate:"required"`
-	Weight   int    `json:"weight" validate:"required"`
-	CourseId string `json:"courseId" validate:"required"`
+	Name     string  `json:"name" validate:"required"`
+	Weight   float64 `json:"weight" validate:"required"`
+	CourseId string  `json:"courseId" validate:"required"`
 }
 
 type UpdateAssignmentGroupPayload struct {
-	Name   string `json:"name" validate:"required"`
-	Weight int    `json:"weight" validate:"required"`
+	Name   string  `json:"name" validate:"required"`
+	Weight float64 `json:"weight" validate:"required"`
 }
